Feature/address dta

+++ src/main/java/pl/orange/NextDoorBook/address/AddressService.java
@@ -1,27 +1,47 @@
 package pl.orange.NextDoorBook.address;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import pl.orange.NextDoorBook.address.DTO.AddressDTO;
+import pl.orange.NextDoorBook.address.DTO.AddressDTOMapper;
+
+import java.util.Optional;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class AddressService {
     private final AddressRepository addressRepository;
+    private final AddressDTOMapper addressDTOMapper;
 
-    public void addAddress(Address address) {
-        addressRepository.addAddress(address);
+    public void addAddress(AddressDTO addressDTO) {
+        addressRepository.addAddress(addressDTOMapper.apply(addressDTO));
     }
 
-    public void deleteAddressById(Long id) {
+    public boolean deleteAddressById(Long id) {
+        if (addressRepository.getAddressById(id).isEmpty()){
+            return false;
+        }
         addressRepository.deleteAddressById(id);
+        return true;
     }
 
-    public void getAddressById(Long id) {
-        addressRepository.getAddressById(id);
+    public Optional<AddressDTO> getAddressById(Long id) {
+        return addressRepository.getAddressById(id).map(addressDTOMapper);
     }
 
-    public void updateAddress(Long id, String cityName, String street, int numberHouse, int zipCode, String district) {
-        addressRepository.updateAddress(id,cityName,street,numberHouse,zipCode,district);
+    public boolean updateAddress(Long id, AddressDTO addressDTO) {
+        if (addressRepository.getAddressById(id).isEmpty()){
+            return false;
+        }
+        addressRepository.updateAddress(
+                id,
+                addressDTO.cityName(),
+                addressDTO.street(),
+                addressDTO.numberHouse(),
+                addressDTO.zipCode(), addressDTO.district());
+        return true;
     }
 
 }
